refactor: create util mixin for course factories with staff
